test: 로또 당첨 테스트 LottoWinMatherTest로 옮김
